Started making scenes
Made a scene tracker file "Scene_list.txt" to keep tack of which scene is which and the functions to navigate to it.
updated the game clock from 30 to 60fps so the typewriter is not as limited.
Updated the Typewriter.py so that text lines can be justified.
Started creating the new scence with templates done for SceneStart, SceneControls and SceneMissionAsteroids.